requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-ignite>=0.4.12
kornia>=0.7.0
wandb>=0.15.12
tqdm>=4.66.1
click>=8.1.7
pydantic>=2.4.2
pydantic-settings>=2.1.0
fastapi>=0.109.0
uvicorn>=0.27.0
accelerate==1.0.0rc1
